test(io): move connection related tests to test_connection.py
